# src/app/docs/swagger-suno-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Suno API Plus Next",
    "description": "Unofficial self-hosted Suno API gateway with account affinity, Studio workflows, Advanced Split, Fast Upload and OpenAI-compatible routes.",
    "version": "2.0.0"
  },
...
  "paths": {
...
    "/api/clip": {
      "get": {
        "summary": "Get clip information based on ID.",
        "description": "Retrieve specific clip information using the provided clip ID as a query parameter.",
        "tags": [
          "default"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "Clip ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "account_affinity",
            "in": "query",
            "required": false,
            "description": "Optional target-account affinity used for deterministic lookup in multi-account mode.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/audio_info"
                }
              }
            }
          },
          "400": {
            "description": "Missing parameter id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Missing parameter id"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/accounts": {
      "get": {
        "summary": "Handle GET /api/admin/accounts.",
        "description": "Implemented by src/app/api/admin/accounts/route.ts. Credentials and account affinity are runtime values; never commit them.",
        "operationId": "get_api_admin_accounts",
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication or API key required."
          },
          "409": {
            "description": "Conflict, active operation, or submission is unknown."
          },
          "429": {
            "description": "Concurrency limit exceeded."
          },
          "503": {
            "description": "Provider or local runtime unavailable."
          }
        }
      },
      "post": {
        "summary": "Handle POST /api/admin/accounts.",
        "description": "Implemented by src/app/api/admin/accounts/route.ts. Credentials and account affinity are runtime values; never commit them.",
        "operationId": "post_api_admin_accounts",
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication or API key required."
          },
          "409": {
            "description": "Conflict, active operation, or submission is unknown."
          },
          "429": {
            "description": "Concurrency limit exceeded."
          },
          "503": {
            "description": "Provider or local runtime unavailable."
          }
        }
      },
      "patch": {
        "summary": "Handle PATCH /api/admin/accounts.",
        "description": "Implemented by src/app/api/admin/accounts/route.ts. Credentials and account affinity are runtime values; never commit them.",
        "operationId": "patch_api_admin_accounts",
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication or API key required."
          },
          "409": {
            "description": "Conflict, active operation, or submission is unknown."
          },
          "429": {
            "description": "Concurrency limit exceeded."
          },
          "503": {
            "description": "Provider or local runtime unavailable."
          }
        }
      },
      "delete": {
        "summary": "Handle DELETE /api/admin/accounts.",
        "description": "Implemented by src/app/api/admin/accounts/route.ts. Credentials and account affinity are runtime values; never commit them.",
        "operationId": "delete_api_admin_accounts",
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication or API key required."
          },
          "409": {
            "description": "Conflict, active operation, or submission is unknown."
          },
          "429": {
            "description": "Concurrency limit exceeded."
          },
          "503": {
            "description": "Provider or local runtime unavailable."
          }
        }
      }
    },
    "/api/admin/accounts/verify": {
      "post": {
        "summary": "Verify an account-pool entry.",
        "description": "Implemented by src/app/api/admin/accounts/verify/route.ts. Credentials and account affinity are runtime values; never commit them.",
        "operationId": "post_api_admin_accounts_verify",
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication or API key required."
          },
          "409": {
            "description": "Conflict, active operation, or submission is unknown."
          },
          "429": {
            "description": "Concurrency limit exceeded."
          },
          "503": {
            "description": "Provider or local runtime unavailable."
          }
        }
      }
    },
    "/api/admin/apikey": {
      "get": {
        "summary": "Handle GET /api/admin/apikey.",
        "description": "Implemented by src/app/api/admin/apikey/route.ts. Credentials and account affinity are runtime values; never commit them.",
        "operationId": "get_api_admin_apikey",
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication or API key required."
          },
          "409": {
            "description": "Conflict, active operation, or submission is unknown."
          },
          "429": {
            "description": "Concurrency limit exceeded."
          },
          "503": {
            "description": "Provider or local runtime unavailable."
          }
        }
      },
      "put": {
        "summary": "Handle PUT /api/admin/apikey.",
        "description": "Implemented by src/app/api/admin/apikey/route.ts. Credentials and account affinity are runtime values; never commit them.",
        "operationId": "put_api_admin_apikey",
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication or API key required."
          },
          "409": {
            "description": "Conflict, active operation, or submission is unknown."
          },
          "429": {
            "description": "Concurrency limit exceeded."
          },
          "503": {
            "description": "Provider or local runtime unavailable."
          }
        }
      }
    },
...
    "/api/admin/captcha": {
      "get": {
        "summary": "Handle GET /api/admin/captcha.",
        "description": "Implemented by src/app/api/admin/captcha/route.ts. Credentials and account affinity are runtime values; never commit them.",
        "operationId": "get_api_admin_captcha",
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication or API key required."
          },
          "409": {
            "description": "Conflict, active operation, or submission is unknown."
          },
          "429": {
            "description": "Concurrency limit exceeded."
          },
          "503": {
            "description": "Provider or local runtime unavailable."
          }
        }
      },
      "put": {
        "summary": "Handle PUT /api/admin/captcha.",
        "description": "Implemented by src/app/api/admin/captcha/route.ts. Credentials and account affinity are runtime values; never commit them.",
        "operationId": "put_api_admin_captcha",
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication or API key required."
          },
          "409": {
            "description": "Conflict, active operation, or submission is unknown."
          },
          "429": {
            "description": "Concurrency limit exceeded."
          },
          "503": {
            "description": "Provider or local runtime unavailable."
          }
        }
      }
    },
...
    "/api/admin/songs": {
      "get": {
        "summary": "Handle GET /api/admin/songs.",
        "description": "Implemented by src/app/api/admin/songs/route.ts. Credentials and account affinity are runtime values; never commit them.",
        "operationId": "get_api_admin_songs",
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication or API key required."
          },
          "409": {
            "description": "Conflict, active operation, or submission is unknown."
          },
          "429": {
            "description": "Concurrency limit exceeded."
          },
          "503": {
            "description": "Provider or local runtime unavailable."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "audio_info": {
        "type": "object",
        "required": [
          "id",
          "title",
          "image_url",
          "lyric",
          "audio_url",
          "video_url",
          "created_at",
          "model_name",
          "status",
          "gpt_description_prompt",
          "prompt",
          "type",
          "tags"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "audio id"
          },
          "title": {
            "type": "string",
            "description": "music title"
          },
          "image_url": {
            "type": "string",
            "description": "music cover image"
          },
          "lyric": {
            "type": "string",
            "description": "music lyric"
          },
          "audio_url": {
            "type": "string",
            "description": "music download url"
          },
          "video_url": {
            "type": "string",
            "description": "Music video download link, can be used to share"
          },
          "created_at": {
            "type": "string",
            "description": "Create time"
          },
          "model_name": {
            "type": "string",
            "description": "Raw Suno provider model identifier used by the generated audio, for example chirp-fenix."
          },
          "status": {
            "type": "string",
            "description": "The generated states include submitted, queue, streaming, complete."
          },
          "gpt_description_prompt": {
            "type": "string",
            "description": "Simple mode on user input prompt, Suno will generate formal prompts, lyrics, etc."
          },
          "prompt": {
            "type": "string",
            "description": "The final prompt for executing the generation task, customized by the user in custom mode, automatically generated by Suno in simple mode."
          },
          "type": {
            "type": "string",
            "description": "Type"
          },
          "tags": {
            "type": "string",
            "description": "Music genre. User-provided in custom mode, automatically generated by Suno in simple mode."
          }
        },
        "title": "audio_info",
        "description": "Audio Info"
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "AdminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "suno_admin_session"
      }
    }
  },
...
}